Move loadSources and graph command functionality into the SpecmaticConfig object, and made source private
